+++ packages/terra-dropdown-button/src/_Dropdown.jsx
@@ -44,6 +44,7 @@ const Dropdown = ({
   requestClose, isOpen, targetRef, children, width, refCallback, buttonRef, getSelectedOptionText, ...customProps
 }) => {
   const buttonFocused = useRef(false);
+  const { menuId, buttonId } = customProps;
   useEffect(() => {
     // added this change to bring focus back to button when dropdown list is closed.
     if (buttonFocused.current && buttonRef) {
@@ -76,8 +77,8 @@ const Dropdown = ({
             width={width}
             refCallback={refCallback}
             getSelectedOptionText={getSelectedOptionText}
-            menuId={customProps.menuId}
-            buttonId={customProps.buttonId}
+            menuId={menuId}
+            buttonId={buttonId}
           >
             {children}
           </DropdownList>